Handle MRI's Errno::EALREADY in connect_nonblock. Fixes #138
In certain high-load situations in RubyDNS (thousands of connections), it
has been observed that `Socket#connect_nonblock` sometimes fails with
`Errno::EALREADY`: a connection is already in progress for the specified
socket. In this case I believe it is prudent that we wait for the socket to
become writable and then attempt to `connect_nonblock` again, at which
point the connection should either fail or give `Errno::EISCONN`.